Move config command behind a factory
