+++ guet/main.py
@@ -1,6 +1,6 @@
 import sys
 
-from guet.commands.config import ConfigSetCommand
+from guet.commands.config.command import ConfigSetCommand
 from guet.commands.get.get_factory import GetCommandFactory
 from guet.commands.init_required_decorator import InitRequiredDecorator
 from guet.factory import CommandFactory
@@ -9,6 +9,7 @@ from guet.commands.addcommitter.factory import AddCommitterFactory
 from guet.commands.init import InitDataSourceCommand
 from guet.commands.setcommitters import SetCommittersCommand
 from guet.commands.start import StartCommand
+from guet.commands.config.factory import ConfigCommandFactory
 
 
 def _command_builder_map():
@@ -17,7 +18,7 @@ def _command_builder_map():
     command_builder_map['init'] = InitDataSourceCommand
     command_builder_map['set'] = SetCommittersCommand
     command_builder_map['start'] = StartCommand
-    command_builder_map['config'] = ConfigSetCommand
+    command_builder_map['config'] = ConfigCommandFactory()
     command_builder_map['get'] = InitRequiredDecorator(GetCommandFactory())
     return command_builder_map
